Moved package verification code function to utils
