+++ v0/builder/builder2v1/build_package.go
@@ -34,7 +34,7 @@ func BuildPackageSection2_1(packageName string, dirRoot string, pathsIgnore []st
 	}
 
 	// get the verification code
-	code, err := spdx.GetVerificationCode2_1(files, "")
+	code, err := utils.GetVerificationCode2_1(files, "")
 	if err != nil {
 		return nil, err
 	}
